refactor: extract screenshot module from bridge
